Question 1.   "Sadguru's Amrit-Tulya - Tea Shop" needs you to test their website that helps track the different items they have for sale.
The functional requirement includes Frontend and Backend specifications.

Frontend Requirements:

Create a page with a component at the top that can be used to add a new item to the inventory. The component should require a name, description, and price for the new item. You can upload a picture of the item as well. Below the component that adds new items to the inventory, display a list containing the name and price of each item in the inventory. If your app supports image uploads, you can display a thumbnail here as well. Each item in the list should have a button to remove it from the inventory. Clicking on an item should bring the user to a new page with all the information about the selected item (name, description, price, and optionally photo).

All API calls should be asynchronous and not cause any full-page reloads. The only full-page reloads the user should see are those caused by navigating between the two pages of the app

Regarding visual styles, the only requirement is that the content of both pages should be both horizontally and vertically centered.  
Backend Requirements:

All of the functions of the frontend needs to be served by an API call and the store’s inventory should be persisted across restarts of the backend process. It will use a SQL database for any persistent data storage needs.
Tracking:

Please do your best to keep track of how long you spend on the functional analysis of each of the backend, frontend functionality, the frontend presentational aspects, and Test Case creation.